% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stranger_text}
\alias{stranger_text}
\title{A transcript of all Stranger Things episodes}
\format{
A [tibble][tibble::tibble-package] with `r nrow(stranger_text)` rows and
  `r ncol(stranger_text)` variables:
\code{utterance}, \code{episode_name}, \code{season}, \code{episode_number},
and \code{character}.
}
\source{
\url{https://subslikescript.com/}
}
\usage{
stranger_text
}
\description{
Each season consists of episodes, each episode is divided into utterances.
One utterance per row.
Where available, the character speaking the utterance is included.
}
\keyword{datasets}
